quiz_creator_progress
def function and while loop was introduces. This program asks the user for questions, possible answers, and correct answer, then will append to a txt file for all the questions and data to be saved. The loop is for asking the user again and again until the user decides to end and exit from the program.