quiz_creator_final
Final Update: This program asks the user for input (questions, choices, and answer), then will be saved to a txt file. For the style This program used a module called colorama, to add a color to the text, for the reader to easily identify the choices from the question, and to the correct answer.